+++ src/Pyro4/nsc.py
@@ -4,6 +4,7 @@ Name server control tool.
 Pyro - Python Remote Objects.  Copyright by Irmen de Jong (irmen@razorvine.net).
 """
 
+from __future__ import print_function
 import sys
 from Pyro4 import naming, errors